requirements.txt
pandas==2.2.3
numpy==2.1.1
yfinance==0.2.43
pandas-datareader==0.10.0
scikit-learn==1.5.2
nltk==3.9.1
textblob==0.18.0.post0
transformers==4.46.3
torch==2.5.0
backtrader==1.9.78.123
streamlit==1.39.0
plotly==5.24.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
